EmailValidationTests and small refactors
