+++ src/main/java/com/project/trackfit/customer/CustomerRetrieveRequestMapper.java
@@ -3,8 +3,10 @@ package com.project.trackfit.customer;
 import org.springframework.stereotype.Service;
 
 import java.util.function.Function;
+
 @Service
-public class CustomerRetrieveRequestMapper  implements Function<Customer,RetrieveCustomerRequest> {
+public class CustomerRetrieveRequestMapper implements Function<Customer,RetrieveCustomerRequest> {
+
     @Override
     public RetrieveCustomerRequest apply(Customer customer) {
         return new RetrieveCustomerRequest(
