+++ src/main/java/com/project/trackfit/trainer/PersonalTrainerServiceImpl.java
@@ -3,7 +3,7 @@ package com.project.trackfit.trainer;
 import com.project.trackfit.core.exception.EmailAlreadyTakenException;
 import com.project.trackfit.core.exception.EmailNotValidException;
 import com.project.trackfit.core.exception.ResourceNotFoundException;
-import com.project.trackfit.core.registration.EmailValidator;
+import com.project.trackfit.core.validation.EmailValidator;
 import lombok.AllArgsConstructor;
 import org.springframework.stereotype.Service;
